pyproject.toml
[project]
name = "telegram-exp"
version = "0.1.0"
description = "A Telegram bot project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.1",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "agno>=0.9.0",
    "sqlalchemy>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
